Initial shot at making `.js` support ESM
This modifies the early ESM implementation to not exclude `.js` files
from ESM. (Note: this still supports `.mjs`). It is created as a
demonstration that `.mjs` should not be a hard requirement to use ESM,
as it introduces needless friction to the existing environment.

**I introduce two command line flags:**

`--esm` - Parse the initial script as ESM
`--cjs` - Parse the initial script as CJS (default)

Use of the `--esm` flag will parse the script as ESM supporting `import`
and allowing the user to import CJS files within the limitations of
Node's implementation. Currently all of core is available for named
imports. Users can require CJS scripts by importing the `require`
function from the `module` core module.

``` js
import { require } from 'module';
```

**Notes on Relative CJS from ESM:**

Supports relative `require` calls when called from an ES Module. This
was temporarily implemented using a hack where the require fn is called,
then the parent caller is matched from the resulting error call stack,
then a new require path is formed from the normalized. This is working
with simple examples, but is obviously not a long-term solution. The fix
is using `import.meta` to get a caller path and matching relative to
that.

Limitations:

- The aforementioned hack for relative path resolution
- `.mjs` will allow arbitrary execution of `node script.js`, this
  approach requires the end user to specify what to start with, which can
  be problematic for CLIs, although they can use `.mjs` in this case to
  make the distinction.

Benefits:

- Separation of module systems, `import` for ESM/or core, `require` for CJS
- Greater compatibility with existing tooling and the web
- Don't need to switch existing projects/tooling to support `.mjs`
- Use of `require` for CJS in ESM, opposed to always using `import`
  despite its numerous limitations with CJS compatibility
- Compatibility with `.mjs`, and still allows the benefits of `.mjs`
  without making it a hard requirement